Move CoR file to subdir
